+++ integrations/room/src/commonMain/kotlin/com/powersync/integrations/room/RoomConnectionPool.kt
@@ -5,6 +5,7 @@ import androidx.room.Transactor
 import androidx.room.execSQL
 import androidx.room.useReaderConnection
 import androidx.room.useWriterConnection
+import androidx.sqlite.SQLiteException
 import androidx.sqlite.SQLiteStatement
 import com.powersync.db.driver.SQLiteConnectionLease
 import com.powersync.db.driver.SQLiteConnectionPool
@@ -73,7 +74,18 @@ public class RoomConnectionPool(
         db.getCoroutineScope().launch {
             val tables = schema.rawTables.map { it.name }.toTypedArray()
             db.invalidationTracker.createFlow(*tables, emitInitialState = false).collect {
-                transferPendingRoomUpdatesToPowerSync()
+                try {
+                    transferPendingRoomUpdatesToPowerSync()
+                } catch (e: SQLiteException) {
+                    // It can happen that we get an update shortly before the database is closed. Since this is
+                    // asynchronous, we'd then be using the database in a closed state, which fails. We handle that by
+                    // stopping the flow collection.
+                    if (e.message == "Connection pool is closed") {
+                        return@collect
+                    }
+
+                    throw e
+                }
             }
         }
     }
@@ -91,7 +103,7 @@ public class RoomConnectionPool(
                 val changed =
                     it.usePrepared("SELECT powersync_update_hooks('get')") { stmt ->
                         check(stmt.step())
-                        json.decodeFromString<Set<String>>(stmt.getText(0))
+                        Json.decodeFromString<Set<String>>(stmt.getText(0))
                     }
 
                 val userTables =
@@ -113,10 +125,6 @@ public class RoomConnectionPool(
 
     override suspend fun close() {
         // Noop, Room database managed independently
-    }
-
-    private companion object {
-        val json = Json {}
     }
 }